+++ src/modules/datepicker/fixtures/angular-pickadate-fixture.html
