+++ src/modules/datepicker/index.js
@@ -1,0 +1,7 @@
+module.exports = (function(){
+  angular.module('form-datepicker', [
+    'schemaForm'
+  ])
+    .directive('pickADate', require('./angular-pickadate'))
+    .config(require('./form-datepicker'))
+})();
